ENYO-566: When deleting a panel, the next panel is displayed briefly before transitioning to expected panel
To eliminate the flashing, I changed it to call setIndex() before
reflowing the controls. This still has the potential to show a flash of
what's behind the the Panels, but I didn't see that in testing. Also,
it doesn't animate to the new location, but I'm not sure if that's even
the desired behavior.