Fix titles in PanelSample and panelMatters
When you delete and add panels, you can end up with more than one panel
with the same title, which is a little confusing. This change ensures
that each new panel gets a new title.

Also changed PanelsSample to use zero-based count, because jumping to
"4" and ending up on "5" was weird.

Bonus: changed "next" and "previous" to update the input to show the
current index.