ENYO-1165: Reduce list-port's initial size
For reasons known only to history, the list-port client control in List
is initially set to a height of 10,000,000px before being dynamically
sized to match the Scroller.

We have a bug report that says this causes the List to display
incorrectly in some cases, in some version of Firefox. They claim
reducing it to 1,000,000px will fix the problem.

Since the list-port only stays that size until the first render, it's
probably okay to make it smaller, as long as it's more than the height
of a reasonable scrren. That should preserve whatever function it
performs.