GF-13581 defect fix
As per the defect the horizontal scrollbar was not appear at the first
launch.
This is because the initial scroll width and container width difference
is zero,
because, initially scroll width is being set to the client area of the
scroller rather than the full width of the scroller element.
Default Scroll Height has been set for Vertical to solve this issue for
vertical list for first launch.
So have added initial default width and height for horizontal as it was
there for vertical.

Enyo-DCO-1.1-Signed-off-by: Ashwini R <ashwini13.r@lge.com>