ENYO-270: Use new Control.retainNode() method in enyo.List
enyo.List has long had a point solution in place to keep the
originating DOM node around during the course of a drag-to-reorder
gesture so that the set of touch events associated with the
gesture don't stop in mid-stream in the case where the list
discards the page containing the node.

We now have a need for the same functionality in other places,
so this functionality is being moved to enyo.Control.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)