ENYO-1097: Assign initialize value to this.finishTransitionInfo
Issue:
The fireTransitionFinish is reference self assigned variable 'this.finishTransitionInfo' when it is revisited with transitionOnComplete set as true. In this case, it throw error like t is undefined in 't.toIndex'.

Fix:
Assign initial value into finishTransitionInfo if it is undefined to avoid initial self reference problem.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com