ENYO-2258: FittableLayout: Remove hiddenReflow / reflowOnShowing
Some logic was recently added to FittableLayout to handle the case
where a layout was reflowed while the associated Control was
hidden; in this case, the reflow was being repeated when the
Control was next shown to ensure accurate measurements for the
children being laid out.

In this change, we remove the logic from FittableLayout in favor
of a simpler and more general solution in enyo/Control. The new
solution works in more cases (such as when an ancestor of the
Control is hidden / shown), avoids creaing a new Observer, avoids
the initial reflow altogether, and applies to all layout kinds.

In order to take full advantage of the solution in enyo/Control,
we short-circuit and return `true` from `reflow()` in the case
where the Control we're doing layout for has a size of 0. This
ensures that we'll get a chance to reflow again when the Control
is next shown.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)